requirements.txt
python-openstackclient>=6.0.0
openstacksdk>=1.0.0
ansible-core>=2.13.0
jmespath>=1.0.1
netaddr>=0.8.0
PyYAML>=6.0
requests>=2.28.0
cryptography>=38.0.0
python-keystoneclient>=5.0.0
